Engine: Use `f64` instead of `f32`, engine-wide.

+++ src/beatrice/app/sample/sample_app.v
@@ -40,21 +40,21 @@ pub fn (mut sample_app SampleApplication) draw(mut graphics renderer.IRenderer) 
 
 		graphics.draw_image(
 			image:    img
-			position: vector.Vector2[f32]{100, 100}
+			position: vector.Vector2[f64]{100, 100}
 			rotation: f32(sample_app.c_engine.time.time) / 10.0
 			origin:   vector.centre
 		)
 	}
 	// Text
 	{
-		graphics.draw_rect(vector.Vector2[f32]{100, 100}, vector.Vector2[f32]{200, 50},
+		graphics.draw_rect(vector.Vector2[f64]{100, 100}, vector.Vector2[f64]{200, 50},
 			renderer.Color.from_rgb[u8](0, 0, 0))
 		sample_app.fonts.draw_text(mut graphics,
 			text:     'Hello, World!'
 			r:        255
 			g:        255
 			b:        255
-			position: vector.Vector2[f32]{100, 150}
+			position: vector.Vector2[f64]{100, 150}
 		)
 
 		sample_app.fonts.draw_text(mut graphics,
@@ -62,19 +62,19 @@ pub fn (mut sample_app SampleApplication) draw(mut graphics renderer.IRenderer) 
 			r:        255
 			g:        255
 			b:        255
-			size:     vector.Vector2[f32]{64, 64}
-			position: vector.Vector2[f32]{700, 600}
+			size:     vector.Vector2[f64]{64, 64}
+			position: vector.Vector2[f64]{700, 600}
 		)
 	}
 	// Geometry
 	{
-		graphics.draw_rect(vector.Vector2[f32]{100, 100}, vector.Vector2[f32]{32, 32},
+		graphics.draw_rect(vector.Vector2[f64]{100, 100}, vector.Vector2[f64]{32, 32},
 			renderer.Color.from_rgb[u8](255, 0, 0))
 
-		graphics.draw_rect_outline(vector.Vector2[f32]{640 - 100 / 2, 360 - 64 / 2}, vector.Vector2[f32]{100, 64},
+		graphics.draw_rect_outline(vector.Vector2[f64]{640 - 100 / 2, 360 - 64 / 2}, vector.Vector2[f64]{100, 64},
 			renderer.Color.from_rgb[u8](255, 255, 255))
 
-		graphics.draw_line(vector.Vector2[f32]{1280, 720}, vector.Vector2[f32]{640, 360},
+		graphics.draw_line(vector.Vector2[f64]{1280, 720}, vector.Vector2[f64]{640, 360},
 			renderer.Color.from_rgb[u8](0, 255, 0))
 	}
 }
